Rename reporter package
