+++ internal/controller/reporter/reporter.go
@@ -1,4 +1,4 @@
-package controller
+package reporter
 
 import (
 	"fmt"
